+++ lib/avatar_rings.dart
@@ -1,4 +1,6 @@
-/// Custom user profile widget, waves-like rings around, badges, status points and many more customizations.
+/// Custom user profile widget, waves-like rings around, badges,
+/// status points and many more customizations.
 library avatar_rings;
 
+export 'src/animated_rings.dart';
 export 'src/avatar_rings.dart';