+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 0.1.2
 
+### Features
+
+- add options to set badge text and color
+- hide or show the rings option
+- option to show the status point and set his color
+- option to set a border for the child
+
+# 0.1.2
+
 - add child padding parameter
 
 # 0.1.1